requirements.txt
google-generativeai
fastf1
streamlit
plotly
pandas
python-dotenv
